% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{grant_base_url}
\alias{grant_base_url}
\title{Simpler Grants.gov API base URL}
\usage{
grant_base_url()
}
\value{
A scalar character URL.
}
\description{
Simpler Grants.gov API base URL
}
\examples{
grant_base_url()
}
