% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{grant_extract_options}
\alias{grant_extract_options}
\title{List extract endpoint options}
\usage{
grant_extract_options()
}
\value{
A named list of documented extract parameters and values.
}
\description{
List extract endpoint options
}
\examples{
grant_extract_options()
}
