% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{grant_search_options}
\alias{grant_search_options}
\title{List opportunity search options}
\usage{
grant_search_options()
}
\value{
A named list of documented search parameters and values.
}
\description{
List opportunity search options
}
\examples{
opts <- grant_search_options()
opts$filters
opts$sort_by
}
